Get information about a dataflow

Get infomration about a dataflow

Usage

get_dataflow_structure(
  req = NULL,
  flowRef,
  id,
  version = "latest",
  agencyID = getOption("sdmxdata.agencyID", "NL1"),
  language = getOption("sdmxdata.language", "nl"),
  cache_dir = tempdir(),
  verbose = getOption("sdmxdata.verbose", FALSE)
)

Arguments

req

A character string with the request to the dataflow

flowRef

A string with the flow reference, in the form of "agencyID,id,version"

id

The id of the dataflow

version

The version of the dataflow, defaults to "latest"

agencyID

The agency ID of the dataflow, defaults to "NL1"

language

The language of the metadata, defaults to "nl"

cache_dir

The directory to cache the data in, set to NULL to disable caching.

verbose

print some information on the console

Value

a list with the dataflow information