Get data from a SDMX API

Get data from an SDMX API

Usage

get_observations(
  req = NULL,
  agencyID,
  id,
  version = "latest",
  ref = NULL,
  startPeriod = NULL,
  endPeriod = NULL,
  filter_on = list(),
  ...,
  language = "en",
  as.data.table = FALSE,
  dim_contents = c("label", "both", "id"),
  attributes_contents = c("label", "id", "both"),
  obs_value_numeric = TRUE,
  raw = FALSE,
  drop_first_columns = !raw,
  cache_dir = tempdir(),
  verbose = getOption("sdmxdata.verbose", FALSE)
)

Arguments

req

An endpoint

agencyID

The agency ID

id

The id of the dataflow

version

The version of the dataflow, default "latest"

ref

The dataflow reference can be used in stead of agencyID, id and version

startPeriod

The start period for which the data should be returned, works only for dataflows with an explicit time dimension.

endPeriod

The end period for which the data should be returned, works only for dataflows with an explicit time dimension.

filter_on

A named list of filters to apply to the data, if not specified or list(), it is the default selection, set to NULL to select all.

...

Additional parameters to pass to the request

language

The language of the metadata

as.data.table

If TRUE return a data.table(), otherwise a data.frame()

dim_contents

The contents of the dimension columns, either "label", "id" or "both"

attributes_contents

The contents of the attribute columns, either "label", "id" or "both"

obs_value_numeric

Should the OBS_VALUE column be coerced to numeric? Default is TRUE

raw

If TRUE return the raw data.frame from the SDMX, otherwise the data.frame is processed

drop_first_columns

Should the first columns be dropped? Default is TRUE (if not raw)

cache_dir

The directory to cache the accompanying meta data, set to NULL to disable caching.

verbose

if TRUE print information about the caching.

Value

data.frame() or data.table::data.table() depending on as.data.table

Examples

if (interactive()) {

  dfs <- list_dataflows()
  ref <- dfs$ref[4]

  obs <- get_observations(
    ref = ref,
    filter_on = list(
      "RegioS" = c("NL01"),
      "Perioden" = c("2024MM01")
    )
  )

  obs
}