Get observations using data-explorer url

Retrieve observations using the url from the data-explorer. The url given will be parsed and translated into a get_observations() call, to get you started. The call will be printed to the console or can be retrieved by settings return_query = TRUE.

Usage

get_observations_by_url(url, return_query = FALSE, verbose = FALSE)

Arguments

url

character, the REST 2.1 url retrieved from an external source

return_query

logical, if TRUE return the query object instead of the observation

verbose

logical, if TRUE print information about the query

Value

when return_query = FALSE a data.frame with the observations, i.e. same output as get_observations(). Otherwise a list with the query object and the parameters

Examples

# retrieve an external url pointing to sdmx data

url <- "https://sdmx-api.beta.cbs.nl/rest/data/NL1.CNVT,DF_85929NED,1.0/..2022JJ00+2023JJ00?dimensionAtObservation=AllDimensions"

obs <- get_observations_by_url(url)
#> Retrieving data using the following statement:
#>   get_observations(endpoint = "https://sdmx-api.beta.cbs.nl/rest", 
#>       agencyID = "NL1.CNVT", id = "DF_85929NED", version = "1.0", 
#>       verbose = FALSE, filter_on = list(Perioden = c("2022JJ00", 
#>       "2023JJ00")))
head(obs)
#>                           Topics        GrootteklasseInOpslagcapaciteit
#> 1 Batterijen; inzet voor opladen   Van meer dan 1 MWh tot en met 10 MWh
#> 2 Batterijen; inzet voor opladen   Van meer dan 1 MWh tot en met 10 MWh
#> 3     Batterijen; eigen verbruik Van meer dan 10 MWh tot en met 100 MWh
#> 4           Batterijen; vermogen  Totaal alle grootteklassen batterijen
#> 5           Batterijen; vermogen  Totaal alle grootteklassen batterijen
#> 6    Batterijen; aantal systemen   Van meer dan 1 MWh tot en met 10 MWh
#>   Perioden OBS_VALUE UNIT_MEASURE UNIT_MULT          OBS_MISSING
#> 1     2022     32072      1000kWh  Eenheden het cijfer is bekend
#> 2     2023     47217      1000kWh  Eenheden het cijfer is bekend
#> 3     2023       979      1000kWh  Eenheden het cijfer is bekend
#> 4     2022       132           MW  Eenheden het cijfer is bekend
#> 5     2023       229           MW  Eenheden het cijfer is bekend
#> 6     2023        31       aantal  Eenheden het cijfer is bekend
#>                                                     DECIMALS      OBS_STATUS
#> 1 het cijfer heeft 0 decimalen of  of de waarde is een tekst Nader voorlopig
#> 2 het cijfer heeft 0 decimalen of  of de waarde is een tekst       Voorlopig
#> 3 het cijfer heeft 0 decimalen of  of de waarde is een tekst       Voorlopig
#> 4 het cijfer heeft 0 decimalen of  of de waarde is een tekst Nader voorlopig
#> 5 het cijfer heeft 0 decimalen of  of de waarde is een tekst       Voorlopig
#> 6 het cijfer heeft 0 decimalen of  of de waarde is een tekst       Voorlopig

query <- get_observations_by_url(url, return_query = TRUE)
query
#> $expr
#> get_observations(endpoint = "https://sdmx-api.beta.cbs.nl/rest", 
#>     agencyID = "NL1.CNVT", id = "DF_85929NED", version = "1.0", 
#>     verbose = FALSE, filter_on = list(Perioden = c("2022JJ00", 
#>     "2023JJ00")))
#> 
#> $args
#> $args$endpoint
#> [1] "https://sdmx-api.beta.cbs.nl/rest"
#> 
#> $args$agencyID
#> [1] "NL1.CNVT"
#> 
#> $args$id
#> [1] "DF_85929NED"
#> 
#> $args$version
#> [1] "1.0"
#> 
#> $args$verbose
#> [1] FALSE
#> 
#> $args$filter_on
#> $args$filter_on$Perioden
#> [1] "2022JJ00" "2023JJ00"
#> 
#> 
#>