List available sdmx dataflows

List the sdmx dataflows from a given endpoint.

Usage

list_dataflows(
  endpoint = NULL,
  agencyID = NULL,
  ...,
  language = NULL,
  cache = TRUE,
  raw = FALSE,
  verbose = getOption("sdmxdata.verbose", FALSE)
)

Arguments

endpoint

A character string or endpoint for a given endpoint.

agencyID

A character string from a given agencyID.

...

saved for future use.

language

The language to use for the text used in the response.

cache

if TRUE cache the list of dataflows

raw

If TRUE return the raw data from the SDMX, otherwise the data is processed.

verbose

if TRUE print information about the dataflows.

Value

a data.frame with available dataflows

Examples

dfs <- list_dataflows()

head(dfs)
#>   agencyID             id version
#> 1      NL1 DF_TESTSET_X01     1.0
#> 2      NL1 DF_TESTSET_X02     1.0
#> 3 NL1.CNVT    DF_24103NED     1.0
#> 4 NL1.CNVT    DF_37230ned     1.0
#> 5 NL1.CNVT       DF_37852     1.0
#> 6 NL1.CNVT    DF_50012NED     1.0
#>                                                                       name
#> 1                                TESTSET (dataflow, only use for TESTSET!)
#> 2                                    TESTSET (dataflow, only for TESTSET!)
#> 3                               Mbo; deelnemers, leerweg, opleiding, regio
#> 4                                  Bevolkingsontwikkeling; regio per maand
#> 5 Gezondheid, leefstijl, zorggebruik en -aanbod, doodsoorzaken; vanaf 1900
#> 6                              Geneesmiddelen; kosten en gebruik 2010-2015
#>                      description                       ref
#> 1                           <NA>   NL1:DF_TESTSET_X01(1.0)
#> 2 for testing .Stat applications   NL1:DF_TESTSET_X02(1.0)
#> 3                           <NA> NL1.CNVT:DF_24103NED(1.0)
#> 4                           <NA> NL1.CNVT:DF_37230ned(1.0)
#> 5                           <NA>    NL1.CNVT:DF_37852(1.0)
#> 6                           <NA> NL1.CNVT:DF_50012NED(1.0)