Base class for SDMX endpoints

Create a:

a known endpoint from endpoints use get_endpoint().

unlisted SDMX 2.1 REST endpoint use sdmx_endpoint().

a new SDMX endpoint use SDMXEndpoint$new().

Base class for SDMX endpoints

See also

Public fields

req

httr2::request() object, can be adjusted to add headers or proxy settings

version

character, the SDMX version of the endpoint

id

character, the id of the endpoint

name

character, the name of the endpoint

url

character, the url of the endpoint

verbose

logical, if TRUE print information the connection

cache_dir

character, the directory to cache the metadata in

language

character, the language to use for the text used in the response

Methods

Public methods

SDMXEndpoint$new()

SDMXEndpoint$clone()

Method new()

create a new SDMXEndpoint object, see also sdmx_endpoint()

Usage

SDMXEndpoint$new(
  url,
  id = gsub("\\W+", "_", url),
  name = url,
  language = NULL,
  cache_dir = file.path(tempdir(), "sdmxdata", id),
  verbose = getOption("sdmxdata.verbose", FALSE)
)

Arguments

url

character, the url of the endpoint

id

character, the id of the endpoint

name

character, the name of the endpoint

language

character, the language to use for the text used in the response

cache_dir

character, the directory to cache the metadata in

verbose

logical, if TRUE print information on the connection

Method clone()

The objects of this class are cloneable with this method.

Usage

SDMXEndpoint$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.